chore: 폴더 분리 및 createQueryKeyStore로 변경
