+++ jest.config.ts
@@ -10,6 +10,7 @@ const config: Config = {
   moduleNameMapper: {
     "\\.(gif|ttf|eot|svg|png)$": "<rootDir>/test/fileMock.ts",
     "^@/(.*)$": "<rootDir>/src/$1", // @ 경로 인식하기 위해 추가
+    "^@features/(.*)$": "<rootDir>/src/features/$1",
   },
   transformIgnorePatterns: ["node_modules/(?!(your-module)/)"],
 };
