--- src/App.tsx
@@ -1,8 +1,0 @@
-import { RouterProvider } from "react-router-dom";
-import router from "./router";
-
-function App() {
-  return <RouterProvider router={router} />;
-}
-
-export default App;
