+++ src/app/App.tsx
@@ -1,0 +1,8 @@
+import { RouterProvider } from "react-router-dom";
+import router from "@/app/router";
+
+function App() {
+  return <RouterProvider router={router} />;
+}
+
+export default App;
