+++ src/features/auth/types.ts
@@ -1,0 +1,35 @@
+export type UserDto = {
+  id: number;
+  name: string;
+  email: string;
+  bio: null | string;
+  avatar: null | string;
+  createdAt: string;
+  updatedAt: string;
+};
+
+export interface SigninRequest {
+  email: string;
+  password: string;
+}
+
+export interface SigninResponse {
+  data: {
+    id: number;
+    name: string;
+    accessToken: string;
+    refreshToken: string;
+  };
+}
+
+export interface MyInfoResponse {
+  data: {
+    id: number;
+    name: string;
+    email: string;
+    bio: null | string;
+    avatar: null | string;
+    createdAt: string;
+    updatedAt: string;
+  };
+}
