+++ src/pages/lps.tsx
@@ -1,0 +1,9 @@
+import LpList from "@features/lps/components/lp-list";
+
+export default function LpsPage() {
+  return (
+    <div className={"flex flex-col justify-center items-center py-10"}>
+      <LpList />
+    </div>
+  );
+}
