--- src/pages/lps/index.tsx
@@ -1,9 +1,0 @@
-import LpList from "@/components/LpList";
-
-export default function LpsPage() {
-  return (
-    <div className={"flex flex-col justify-center items-center py-10"}>
-      <LpList />
-    </div>
-  );
-}
