+++ src/pages/my.tsx
@@ -1,0 +1,5 @@
+import MyInfo from "@features/auth/components/my-info";
+
+export default function MyPage() {
+  return <MyInfo />;
+}
