--- src/pages/my/index.tsx
@@ -1,5 +1,0 @@
-import MyInfo from "@/components/MyInfo";
-
-export default function MyPage() {
-  return <MyInfo />;
-}
