+++ src/shared/utils/email.ts
